Tarefa1_2022li1g095

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido

Arguments

:: Mapa
Mapa a avaliar

-> Bool
Validade do Mapa

A função 'mapaValido’ verifica a validade de um mapa. Caso o resultado desta seja True, este será válido. Se for False, então não será válido.

Para a elaboração desta função foram necessárias as funções auxiliares terrenoProprio, riosValido, obstaculoComprimento, larguraComprimento e terrenoValida, cujo propósito de cada uma será explicado ao longo deste trabalho.

Assim, a obtenção de um mapa válido, só é possível se for obtido um resultado True em cada função auxiliar, ou seja, este necessita de respeitar as condições que estas exigem.

Reçalva-se que, para exemplos fornecidos para as funções auxiliares, a validade de cada uma é independente das demais.

De seguida, irá ser demonstrado um exemplo de mapa válido, ou seja, que vai de acordo com todas auxiliares, e um mapa inválido, ou seja, que não vai de acordo com uma ou mais auxiliares.

Exemplo de um mapa válido

>>> mapaValido (Mapa 5 [(Estrada 3, [Nenhum, Carro, Carro, Carro, Nenhum]),
                        (Rio 4, [Nenhum, Tronco, Nenhum, Nenhum, Tronco]), 
                        (Rio (-3), [Tronco, Tronco, Nenhum, Nenhum, Nenhum]), 
                        (Rio 3, [Tronco, Nenhum, Nenhum, Tronco, Tronco]), 
                        (Relva, [Nenhum, Nenhum, Arvore, Arvore, Arvore])])))
True
Este mapa respeita todas as condições estabelecidas nas funções auxiliares, logo, o mapa é válido.

Exemplo de um mapa inválido

>>> mapaValido (Mapa 3 [(Estrada 3, [Nenhum, Tronco, Tronco]),
                        (Rio 4, [Nenhum, Tronco, Nenhum]), 
                        (Rio (-3), [Tronco, Tronco, Nenhum]), 
                        (Rio 3, [Tronco, Nenhum, Nenhum]), 
                        (Relva, [Nenhum, Nenhum, Arvore])])))
False
Nesta situação exemplo, não foi respeitada a auxiliar 'terrenoProprio', pois podemos observar obstáculos em terrenos impróprios. 
Por outras palavras, não deveriam existir troncos em estradas, tal acontecimento classifica o mapa como inválido.

terrenoProprio

Arguments

:: Mapa
Mapa a avaliar

-> Bool
Verificação da existência de obstáculos nos terrenos que lhes são devididamente destinados

A função terrenoProprio verifica se não existem obstáculos em terrenos impróprios, e.g. troncos em estradas ou relvas, árvores em rios ou estradas, etc.

Exemplo de um mapa com terreno válido

>>> terrenoProprio (Mapa 5 [(Estrada 3, [Nenhum, Carro]),
                            (Rio 4,[Nenhum, Tronco]), 
                            (Relva, [Nenhum, Arvore, Arvore, Nenhum])])
True
Este mapa possui carros apenas em estradas, troncos apenas em rios e árvores apenas na relva.
Logo cada obstáculo está no respetivo terreno, o que torna o mapa válido para esta função. 

Exemplo de um mapa com terreno inválido

>>> terrenoProprio (Mapa 5 [(Estrada 3, [Nenhum, Carro]),
                            (Rio 4,[Arvore, Tronco]), 
                            (Relva, [Nenhum, Arvore, Arvore, Nenhum])])
False
Este mapa possui uma árvore em um rio, ou seja, um obstáculo num terreno impróprio. 
Logo, o mapa é inválido para esta função.

riosValido

Arguments

:: Mapa
Mapa a avaliar

-> Bool
Verificação de rios contíguos possuirem sempre direções opostas.

A função riosValido verifica se rios contíguos têm direcões opostas.

Exemplo de um mapa com uma sequência de rios válida

>>> riosValido (Mapa 5 [(Rio 4, [Nenhum, Carro]),
                        (Rio (-4),[Nenhum, Tronco]), 
                        (Relva, [Nenhum, Arvore, Arvore, Tronco]), 
                        (Rio (-1),[Nenhum, Tronco]), 
                        (Rio 2,[Tronco, Nenhum])])
True                        
Todos os rios contíguos, presentes neste mapa, possuem direções opostas. 
Logo, o mapa é válido para esta função.

Exemplo de um mapa com uma sequência de rios válida

>>> riosValido (Mapa 5 [(Rio 4, [Nenhum, Carro]),
                        (Rio (-4),[Nenhum, Tronco]), 
                        (Relva, [Nenhum, Arvore, Arvore, Tronco]), 
                        (Rio (-1),[Nenhum, Tronco]), 
                        (Rio (-2),[Tronco, Nenhum])])
False      
Podemos observar rios contíguos, neste mapa, com a mesma direção, pois possuem o mesmo sinal (negativo). 
Logo, o mapa é inválido para esta função.

obstaculoComprimento

Arguments

:: Mapa
Mapa a avaliar

-> Bool
Verificação do comprimento máximo dos troncos e dos carros, tal como a existência de obstáculos do tipo nenhum em cada linha

A função obstaculoComprimento verifica se o comprimento máximo dos troncos é 5 unidades, se o comprimento máximo dos carros é 3 unidades e se em qualquer linha existe, no mínimo, um obstáculo do tipo Nenhum.

Exemplo de um mapa com sequências de obstáculos com comprimentos adequados e que possuem o obstáculo Nenhum

>>> obstaculoComprimento (Mapa 5 [(Estrada 4, [Nenhum, Carro]),
                                 (Rio (-4), [Nenhum, Tronco, Tronco, Tronco, Tronco, Tronco]), 
                                 (Relva, [Nenhum, Arvore, Arvore, Tronco]), 
                                 (Rio (-1),[Nenhum, Tronco]), 
                                 (Rio (-2),[Tronco, Nenhum])])
True                        
Este mapa têm sequências de obstáculos que não excedem os valores estabelecidos, tal como possui em cada linha o obstáculo Nenhum.
Logo, o mapa é válido para esta função.

Exemplos de mapas com sequências de obstáculos com comprimentos inadequados e que não possuem o obstáculo Nenhum, respetivamente

>>> obstaculoComprimento (Mapa 5 [(Rio 4, [Nenhum, Carro]),
                                  (Rio (-4),[Nenhum, Tronco,Tronco, Tronco, Tronco, Tronco, Tronco]), 
                                  (Relva, [Nenhum, Arvore, Arvore, Tronco]), 
                                  (Rio (-1),[Nenhum, Tronco]), 
                                  (Rio (-2),[Tronco, Nenhum])])
False           
Este mapa possui uma sequência de troncos que ultrapassada o comprimento máximo estabelecido. 
Logo, o mapa é inválido para esta função.

>>> obstaculoComprimento (Mapa 5 [(Rio 4, [Nenhum, Carro]),
                                  (Rio (-4),[Nenhum, Tronco,Tronco, Tronco, Tronco, Tronco, Tronco]), 
                                  (Relva, [Nenhum, Arvore, Arvore, Tronco]), 
                                  (Rio (-1),[Nenhum, Tronco]), 
                                  (Rio (-2),[Tronco, Tronco])])
False
Este mapa possui uma sequência de obstáculos sem o obstáculo Nenhum, desrespeitando um critério estabelecido. 
Logo, o mapa é inválido para esta função.

count_maxSeq :: Eq a => a -> [a] -> Int

Para a elaboração da função obstaculoComprimento, foi necessária a elaboração de uma função euxiliar, count_maxSeq, que conta a maior sequência de um determiando elemento numa lista.

larguraComprimento

Arguments

:: Mapa
Mapa a avaliar

-> Bool
Verificação do comprimento da lista de obstáculos de cada linha corresponder à largura do mapa

A função larguraComprimento verifica se o comprimento da lista de obstáculos de cada linha corresponde exatamente à largura do mapa.

Exemplo de um mapa cujo comprimento da lista de obstáculos é igual à largura

>>> larguraComprimento (Mapa 5 [(Rio 4, [Nenhum, Carro, Carro, Nenhum, Carro]),
                                (Rio (-4),[Nenhum, Tronco,Tronco, Tronco, Tronco])])
True
Este mapa possui uma largura de 5 unidades e sequências com o mesmo número de obstáculos. 
Logo, o mapa é válido para esta função.

Exemplo de um mapa cujo comprimento da lista de obstáculos não é igual à largura

>>> larguraComprimento (Mapa 5 [(Rio 4, [Nenhum, Carro, Nenhum, Carro]),
                                 (Rio (-4),[Nenhum, Tronco,Tronco, Tronco])])
False
Este mapa possui uma largura de 5 unidades e um número inferior de obstáculos na mesma linha. 
Logo, o mapa é válido é inválido para esta função.

terrenoValida

Arguments

:: Mapa
Mapa a avaliar

-> Bool
Verificação da não existência de mais do que 4 rios, nem 5 estradas ou relvas contíguos.

A função terrenoValida verifica se, contiguamente, não existem mais do que 4 rios, nem 5 estradas ou relvas.

Exemplo de um mapa com uma sequência de terrenos válida

>>> terrenoValida (Mapa 5 [(Rio 4, [Nenhum, Carro, Carro, Nenhum, Carro]),
                           (Rio (-4),[Nenhum, Tronco,Tronco, Tronco]), 
                           (Rio 4, [Nenhum, Carro, Carro, Nenhum, Carro]), 
                           (Rio (-4),[Nenhum, Tronco, Tronco, Tronco]),
                           (Estrada 2 [Carro, Carro, Carro, Nenhum])])
True
Este mapa possui uma quantidade de sequẽncias de terrenos contíguos que não excedem os valores estabelecidos. 
Logo, o mapa é válido para esta função.

Exemplo de um mapa com uma sequência de terrenos inválida

>>> terrenoValida (Mapa 5 [(Estrada 3, [Nenhum, Carro]),
                           (Rio 4,[Nenhum, Tronco]), 
                          (Rio 2, [Nenhum, Arvore, Arvore, Nenhum]), 
                          (Rio 2, [Nenhum, Tronco]), 
                          (Rio 4, [Tronco, Tronco, Nenhum]), 
                          (Rio 2, [Tronco, Nenhum])])
False
Este mapa possui uma quantidade de rios contíguos que excedem os valores estabelecidos. 
Logo, o mapa é inválido para esta função.

contaRios :: Mapa -> Int -> Int -> Int

Para a elaboração da função terrenoValida, foi necessário o desenvolvimento da função auxiliar contaRios, que conta a maior sequência de rios seguidos numa lista.

contaEstradas :: Mapa -> Int -> Int -> Int

Para a elaboração da função terrenoValida, foi necessário o desenvolvimento da função auxiliar contaEstradas, que conta a maior sequência de estradas seguidas numa lista.

contaRelva :: Mapa -> Int -> Int -> Int

Para a elaboração da função terrenoValida, foi necessário o desenvolvimento da função auxiliar contaRelva, que conta a maior sequência de relvados seguidos numa lista.