Tarefa2_2022li1g095

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa

Arguments

:: Mapa
Mapa ao qual se vai acrescentar uma linha

-> Int
Inteiro aleatório que fornece pseudo-aleatoriedade à geração da nova linha

-> Mapa
Mapa gerado

A função estendeMapa cria uma nova linha válida no topo de um mapa.

Para a elaboração desta função foram necessárias as funções auxiliares atualizatualizaVelocidadeEArvoresaVelocidade, constroiListaObstaculos, proximosTerrenosValidos, proximosObstaculosValidos e validaInicioLista, cujo propósito de cada uma será explicado ao longo deste trabalho.

Assim, observa-se que o novo mapa obtido, respeita as regras de todas as funções auxiliares referidas.

De seguida, irá ser demonstrado um exemplo de um mapa que foi amplificado, de acordo com as funções auxiliares.

Propriedade

Assume-se que o mapa fornecido como parâmetro é valido.

Esta função não permite casos em que a passagem do jogador entre árvores seja impossível. Ou seja, entre 2 linhas contíguas de relva garante-se sempre

uma alternativa de passagem.

Exemplo de um mapa, ao qual foi acrescentada uma linha

>>> estendeMapa (Mapa 6 [(Rio 3, [Nenhum, Tronco, Tronco, Tronco, Nenhum, Tronco]),
                         (Estrada 2, [Carro, Carro, Nenhum, Carro, Nenhum, Nenhum]), 
                         (Relva, [Arvore, Arvore, Nenhum, Nenhum, Arvore, Nenhum])]) 20
Mapa 6 [(Estrada (-4), [Nenhum, Carro, Nenhum, Nenhum, Nenhum, Carro]),
        (Rio 3, [Nenhum, Tronco, Tronco, Tronco, Nenhum, Tronco]),
        (Estrada 2, [Carro, Carro, Nenhum, Carro, Nenhum, Nenhum]),
        (Relva, [Arvore, Arvore, Nenhum, Nenhum, Arvore, Nenhum])]

atualizaVelocidadeEArvores

Arguments

:: Mapa
Mapa com um terreno com uma velocidade por atribuir

-> Int
Inteiro aleatório que fornece pseudo-aleatoriedade à geração da nova linha

-> Mapa
Mapa gerado com a velocidade atualizada

A função atualizaVelocidadeEArvores permite que seja definida uma velocidade para os novos terrenos gerados.

Propriedades

Convencionou-se, valores de velocidade compreendidos entre -5 e 5, excepto 0.

Para isto, utilizou-se o parâmetro de entrada da função principal estendemapa para gerar números aleatórios no gerador mkStdGen. De seguida, recorreu-se à função randomR para garantir um intervalo de valores desejados. Foi necessário recorrer a este gerador, tal como à função referida para que a função atualizaVelocidadeEArvores consiga fornecer alguma pseudo-aleatoriedade.

Para rios contíguos atribui-se velocidades com sinais opostos, com o objetivo destes possuirem direções opostas.

No caso do terreno ser relva, a velocidade não é atualizada.

Exemplo

>>> atualizaVelocidadeEArvores (Mapa 3 [(Rio 0,[Tronco,Nenhum,Tronco]),
                                (Rio (-2),[Tronco, Tronco, Nenhum]),
                                (Relva, [Arvore,Arvore,Nenhum]),
                                (Estrada 1,[Nenhum,Carro,Carro])]) 89)
Mapa 3 [(Rio 3, [Tronco, Nenhum, Tronco]),
        (Rio (-2), [Tronco, Tronco, Nenhum]),
        (Relva, [Arvore, Arvore, Nenhum]),
        (Estrada 1, [Nenhum, Carro, Carro])]
Como podemos observar, ao rio, que não possuia velocidade, foi atribuida uma velocidade aleatória, compreendida entre -5 a 5, sendo
esse valor 3.

garantePassagemEntreArvores

Arguments

:: [Obstaculo]
Primeira linha de obstáculos

-> [Obstaculo]
Segunda linha de obstáculos

-> Int
Largura do mapa

-> Int
Inteiro aleatório que fornece pseudo-aleatoriedade à geração da nova lista

-> [Obstaculo]
Lista final

Para a elaboração da função atualizaVelocidadeEArvores foi necessário o desenvolvimento da função auxiliar garantePassagemEntreArvores. Esta permite a construção de uma nova linha que garanta a passagem do jogador num terreno com relva. Caso seja possível o jogador atravessar os terrenos com relva, então a função manterá a lista dada como parâmetro.

Exemplo

>>> garantePassagemEntreArvores [Arvore,Arvore,Nenhum]
                                [Nenhum,Arvore,Arvore] 3 10
[Nenhum,Arvore,Nenhum]
Nesta situação, como a passagem era impossível, esta função reconstruiu a primeira linha e forneceu uma alternativa para esta.

validaPassagemEntreArvores

Arguments

:: [Obstaculo]
Primeira lista de obstáculos da relva

-> [Obstaculo]
Segunda lista de obstáculos da relva

-> Bool
Valida a passagem

Para a elaboração da função garantePassagemEntreArvores foi necessário o desenvolvimento da função auxiliar validaPassagemEntreArvores. Esta recebe 2 listas de obstáculos do terreno relva e avalia se existe uma passagem possível para o Jogador.

Exemplos

>>> validaPassagemEntreArvores [Arvore,Arvore,Nenhum] [Nenhum,Arvore,Arvore]
False
Nesta situação, não existe passagem.

>>> validaPassagemEntreArvores [Nenhum,Arvore,Nenhum] [Nenhum,Arvore,Arvore]
True
Nesta situação, existe passagem.

randomRExcetoNum :: Int -> Int -> Int -> Int -> Int

Para a elaboração da função atualizaVelocidadeEArvores, foi necessária a formulação da função auxiliar randomRExcetoNum. Esta calcula um número aleatório entre um intervalo, excluindo o número passado como parâmetro. No caso do cálculo da velocidade em ros e estradas pretende-se excluir o número 0 do intervalo -5 e 5. Para que isto seja possível, utilizou-se a função randomR, que caso o resultado obtido seja 0, esta é invocada novamente somando 100 ao parâmetro do gerador mkStdGen.

constroiListaObstaculos

Arguments

:: Int
Largura

-> Int
Inteiro aleatório que fornece pseudo-aleatoriedade à geração da nova linha

-> (Terreno, [Obstaculo])
Terreno da linha a construir

-> Int
Condição de paragem

-> [Obstaculo]
Lista de obstáculos gerada

A função constroiListaObstaculos tem como onjetivo gerar uma nova lista válida de obstáculos, indo de acordo com a primeira tarefa.

Exemplo

>>> constroiListaObstaculos 3 60 (Relva, []) 3
[Arvore,Nenhum,Nenhum]

proximosTerrenosValidos

Arguments

:: Mapa
Mapa ao qual podem ser adicionados terrenos

-> [Terreno]
Lista de possíveis terrenos válidos a acrescentar ao topo do mapa

A função proximosTerrenosValidos fornece uma lista de possibilidades de terrenos válidos que podem surgir.

Propriedades

Definiu-se que a velocidade dos terrenos gerados é 0, pois esta função auxiliar não é responsável por gerar velocidades.

Definiu-se que, contiguamente, não devem existir mais do que 4 rios, nem 5 estradas ou relvas, respeitando a primeira tarefa.

Exemplos de terrenos válidos que podem surgir

>>> proximosTerrenosValidos (Mapa 3 [(Estrada 3, [Nenhum, Tronco, Tronco]),
                                     (Rio 4, [Nenhum, Tronco, Nenhum]), 
                                     (Rio (-3), [Tronco, Tronco, Nenhum]), 
                                     (Rio 3, [Tronco, Nenhum, Nenhum]), 
                                     (Relva, [Nenhum, Nenhum, Arvore])])
[Rio 0, Estrada 0, Relva]

>>> proximosTerrenosValidos (Mapa 3 [(Rio 3, [Nenhum, Tronco, Tronco]),
                                     (Rio 3, [Nenhum, Tronco, Nenhum]), 
                                     (Rio (-3), [Tronco, Tronco, Nenhum]), 
                                     (Rio (-3), [Tronco , Nenhum, Nenhum]), 
                                     (Rio 3, [Nenhum, Nenhum, Tronco]), 
                                     (Relva, [Arvore, Nenhum, Nenhum])])
[Estrada 0, Relva]

proximosObstaculosValidos

Arguments

:: Int
Largura do mapa

-> (Terreno, [Obstaculo])
Linha incompleta do mapa

-> Int
Posicação da lista em construção

-> [Obstaculo]
Lista de obstáculos válidos para continuar uma dada linha do mapa

A função proximosObstaculosValidos fornece uma lista de possibilidades de obstáculos válidos que podem surgir.

Propriedade

Pressupõe-se que os obstáculos são acrescentados no início da lista de obstáculos, ou seja, à esquerda.

Não são geradas listas de obstáculos totalmente preenchidas por arvores, troncos, carros ou por nenhum.

Se um tronco está a atrevessar um wormhole e o seu comprimento já for máximo, não se pode acrescer mais unidades do tipo tronco ao início da lista, apesar do tronco não aprensetar as 5 unidades juntas.

O comprimento de uma sequência de árvores não é validado por esta função.

Exemplos de obstáculos válidos que podem surgir

>>> proximosObstaculosValidos 5 (Estrada 3, [Carro, Carro, Nenhum, Carro])
[Nenhum]

>>> proximosObstaculosValidos 5 (Estrada 3, [Nenhum, Carro, Nenhum, Carro])
[Nenhum, Carro]

validaInicioLista :: Obstaculo -> Int -> [Obstaculo] -> Bool

Para a elaboração da função proximosObstaculosValidos, foi necessário o desenvolvimento da função auxiliar validaInicioLista. Esta permite, na construção de uma lista da direita para a esquerda, impedir situações em que sequências seguidas excedam o comprimento máximo definido para os carros e para os troncos.

acertaInicioListaRio :: [Obstaculo] -> [Obstaculo]

Para a elaboração da função constroiListaObstaculos, foi necessário o desenvolvimento da função auxiliar acertaInicioListaRio. Esta permite na construção de uma lista, da direita para a esquerda, impedir que troncos, de comprimento máximo, ao atravessar wormholes aumentem o seu tamanho.

acertaInicioListaEstrada :: [Obstaculo] -> [Obstaculo]

Para a elaboração da função constroiListaObstaculos, foi necessário o desenvolvimento da função auxiliar acertaInicioListaEstrada. Esta permite na construção de uma lista, da direita para a esquerda, impedir que carros, de comprimento máximo, ao atravessar wormholes aumentem o seu tamanho.